feat(library): 完善删除书籍功能
- 在BookProgressDao中添加删除阅读进度的方法
- 更新DeleteBookUseCase注入进度和历史仓库依赖
- 在删除书籍时同步删除对应阅读进度和历史记录
- 修复HistoryDao中bookId参数类型从Int改为Long
- 在ProgressRepository和ReadHistoryRepository中添加删除方法
- 实现ProgressRepositoryImpl和ReadHistoryRepositoryImpl的删除功能

+++ data/src/main/java/io/github/lycosmic/data/local/dao/HistoryDao.kt
@@ -43,7 +43,7 @@ interface HistoryDao {
      * 删除指定书籍 id 的阅读记录
      */
     @Query("DELETE FROM ${HistoryEntity.TABLE_NAME} WHERE ${HistoryEntity.COL_BOOK_ID} = :bookId")
-    suspend fun deleteHistoryByBookId(bookId: Int): Int
+    suspend fun deleteHistoryByBookId(bookId: Long): Int
 
     /**
      * 删除阅读记录
